♻️ refactor :  investments에 있는 footer 데이터를 따로 뻄
