+++ OneCoin/Client/src/utills/constants/footer.ts
@@ -1,0 +1,15 @@
+export const FOOTER_LIST = {
+	repository: 'https://github.com/codestates-seb/seb40_main_020',
+	copyright: '1Coin © 2022 All rights reserved.',
+	fe: [
+		{ name: '서재홍', repository: 'https://github.com/jaehongg' },
+		{ name: '김민상', repository: 'https://github.com/minsang98' },
+		{ name: '민다영', repository: 'https://github.com/MINDA01' },
+		{ name: '유현규', repository: 'https://github.com/korean-davinci' },
+	],
+	be: [
+		{ name: '김기홍', repository: 'https://github.com/broaden-horizon' },
+		{ name: '김지환', repository: 'https://github.com/kjh42447' },
+		{ name: '남기범', repository: 'https://github.com/GIVEN53' },
+	],
+};
